8_19

+++ C++/override_and_load/test.cpp
@@ -125,7 +125,7 @@ public:
 class Derived:public Base
 {
 public:
-	void f(int x)
+	void f(float x)
 	{
 		cout<<"Derived::f(float)"<<endl;
 	}
